+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 ### Fixed
+- Don't crash target table when port_list is undefined [#3120](https://github.com/greenbone/gsa/pull/3120)
+
 
 [Unreleased]: https://github.com/greenbone/gsa/compare/v21.4.2...gsa-21.04